fixed dev script, added DB for dev version, moved registerModules() inside client class

+++ src/bot/modules/Player.ts
@@ -53,24 +53,26 @@ let embedTemplate = {
 type Song = {
 	title: string,
 	link: string,
-	repeat: boolean
+	repeat: boolean,
+	duration: string
 }
 type Queue = {
 	[key: string]: Song[] 
 }
+
 export class Player extends BaseModule {
 	private client: IOEClient;
 	private reactions: string[] = ['▶', '⏸', '⏹', '⏭', '🔁', '🔀', '🇯']; // '🔇', '🔉', '🔊'
 	private channels: Map<string, string>;
 	private playerControllMessages: Map<string, Message> = new Map();
 	private player: Map<string, AudioPlayer> = new Map();
-
+	private blockedUsers: Map<string, boolean> = new Map();
 	constructor(client: IOEClient) {
 		super(NAME);
 		this.client = client;
-		
+
 		this.init()
-		this.log('Initialization completed.')
+		
 	}
 	async init() {
 		try {
@@ -86,7 +88,7 @@ export class Player extends BaseModule {
 				}
 
 			})
-			
+			this.log('Initialization completed.')
 			
 		} catch (e) {
 			this.log('Init Error: ', e)
@@ -114,25 +116,49 @@ export class Player extends BaseModule {
 	}
 	async nextSong(guildId: string, force?: boolean) {
 		try {
-			if (db.exists('/queue')) {
-				let queue = db.getObject<Queue>('/queue')
-				if (!queue[guildId] || queue[guildId].length == 0) return;
-				if (!queue[guildId][0].repeat || force) {
-					queue[guildId].shift();
-					db.push('/queue', queue);
-				}
+			let queue = await this.getQueue(guildId)
 				
-				await this.updateControllMessage(guildId);
-				if (queue[guildId].length == 0) {
+				if (!queue[0].repeat || force) {
+					queue.shift();
+					await this.setQueue(guildId, queue)
+				}
+					await this.updateControllMessage(guildId);
+					
+				if (queue.length == 0) {
 					const connection = getVoiceConnection(guildId);
 					if (connection) connection.destroy();
 					return;
 				};
+				
 				await this.play(guildId);
-			}
+			
 		} catch (e) {
 			this.log('Skip error:', e)
 		}
+	}
+	async getQueue(guildId: string) {
+		if (db.exists('/queue')) {
+			let queue = db.getObject<Queue>('/queue')
+			if (!queue[guildId]) {
+				queue[guildId] = []
+				db.push('/queue', queue)
+				return queue[guildId]
+			}
+			return queue[guildId];
+
+		} else {
+			let queue: Queue = {}
+			queue[guildId] = []
+			db.push('/queue', queue)
+			return queue[guildId]
+		}
+	}
+	async setQueue(guildId: string, songs: Song[]) {
+		await this.getQueue(guildId);
+		let queue = db.getObject<Queue>('/queue');
+		queue[guildId] = songs
+		db.push('/queue', queue)
+		return queue[guildId]
 	}
 	async reactionHandler(reaction: MessageReaction | PartialMessageReaction, user: User | PartialUser) {
 		try {
@@ -146,12 +172,15 @@ export class Player extends BaseModule {
 			const player = this.player.get(guildId);
 			if (!player) return;
 			switch (reaction.emoji.name) {
+				// Play
 				case (this.reactions[0]):
 					player.unpause()
 					break
+				// Pause
 				case (this.reactions[1]):
 					player.pause(true)
 					break
+				// Stop
 				case (this.reactions[2]):
 					let connection = getVoiceConnection(guildId);
 					if (connection) connection.destroy();
@@ -164,41 +193,46 @@ export class Player extends BaseModule {
 					}
 					await this.updateControllMessage(msg.guildId)
 					break
+				// Skip
 				case (this.reactions[3]):
 					await this.nextSong(guildId, true);
 					break
+				// Repeat
 				case (this.reactions[4]):
-					let queue = db.getObject<Queue>('/queue')
-					if (!queue[guildId] || queue[guildId].length == 0) break;
-					queue[guildId][0].repeat = !queue[guildId][0].repeat;
-					db.push('/queue', queue);
+					let queue = await this.getQueue(guildId);
+					if (queue.length == 0) break;
+					queue[0].repeat = !queue[0].repeat;
+					await this.setQueue(guildId, queue)
 					await this.updateControllMessage(guildId)
 					break
+				// Shuffle
 				case (this.reactions[5]):
-					if (db.exists('/queue')) {
-						let queue = db.getObject<Queue>('/queue')
-						if (!queue[guildId] || queue[guildId].length <= 1) break;
-						let currentSong = queue[guildId].shift()
-						let shuffled = await this.shuffle(queue[guildId]);
-						shuffled.unshift(currentSong)
-						queue[guildId] = shuffled;
-						db.push('/queue', queue)
-						await this.updateControllMessage(guildId)
-					}
+
+					queue = await this.getQueue(guildId)
+					if (queue.length <= 1) break;
+					let currentSong = queue.shift()
+					let shuffled = await this.shuffle(queue);
+					shuffled.unshift(currentSong)
+						
+					await this.setQueue(guildId, shuffled);
+					await this.updateControllMessage(guildId);
+					
 					break
+				// Join channel
 				case (this.reactions[6]):
-				
+					let guild = await this.client.guilds.fetch(guildId)
+					let newChannelId = (await guild.members.fetch(user.id)).voice.channelId
 					let Oldconnection = getVoiceConnection(guildId);
+					if (Oldconnection.joinConfig.channelId == newChannelId) break;
 					if (Oldconnection) Oldconnection.destroy();
 				
-					console.log(player.state.status)
 					if (player.state.status !== AudioPlayerStatus.Idle) {
 						player.pause(true);
 					};
 				
-					let guild = await this.client.guilds.fetch(guildId)
+					
 					let newConnection = joinVoiceChannel({
-						channelId: (await guild.members.fetch(user.id)).voice.channelId,
+						channelId: newChannelId,
 						guildId: guildId,
 						adapterCreator: guild.voiceAdapterCreator
 					})
@@ -215,83 +249,94 @@ export class Player extends BaseModule {
 	sendChooseMessage(member: GuildMember, channel: TextChannel, tracks: Search.YouTubeSearchResults[]): Promise<number> {
 		return new Promise(async (resolve, reject) => {
 			try {
-				const timeoutId = setTimeout(async () => {
-					await removeMsgAndCollector();
-					resolve(-1)
-				}, 50000)
+				
 		
-				let embed = Object.assign({}, embedTemplate)
+				let embed: typeof embedTemplate = JSON.parse(JSON.stringify(embedTemplate));
 				embed.author.name = "Выберите трек:"
-				let reactions = ['1️⃣', '2️⃣', '3️⃣', '4️⃣', '5️⃣']
-				let cancel = '❌'
+				//let reactions = ['1️⃣', '2️⃣', '3️⃣', '4️⃣', '5️⃣']
+				//let cancel = '❌'
 				let tracksCount = 0;
 				for (let i = 0; i < tracks.length; i++) {
 					if (i >= 5) break;
 					tracksCount += 1;
 					embed.description += `${i + 1}. **[${tracks[i].title}](${tracks[i].link})** \n`;
 				}
-		
-			
+				
+				 
+				embed.fields[0] = {
+					"name": "\u200B",
+					"value": "`С или Cancel чтобы отменить.\n`",
+					"inline": true
+				};
 				let chooseMsg = await channel.send({
 					embeds: [embed]
 				});
-				const reactionCollector = await chooseMsg.createReactionCollector()
-				reactionCollector.on('collect', (r, user) => {
-					if (user.bot) return;
-					if (user.id !== member.user.id) return;
-					const emoji = r.emoji.name
+				const messageCollector = await channel.createMessageCollector();
+				const timeoutId = setTimeout(async () => {
+					await removeMsgAndCollector();
+					resolve(-1)
+				}, 30000)
+				const removeMsgAndCollector = async () => {
+					try {
+						this.blockedUsers.set(member.id, false)
+						clearTimeout(timeoutId);
+						messageCollector.stop()
+						if (!chooseMsg.deleted) {
+							await chooseMsg.delete();
+						}
+					} catch (e) {
+						this.log('Remove choose msg error', e)
+					}
+				}
+				messageCollector.on('collect', (m) => {
+					if (m.member.user.id !== member.user.id) return;
+					if (m.author.bot) return;
+					if(!m.deleted) m.delete()
+				
+					switch (m.content.toLocaleLowerCase()) {
 
-					switch (emoji) {
-						case (reactions[0]):
+						case ('1'):
 							resolve(0)
 							removeMsgAndCollector()
 							break
-						case (reactions[1]):
+
+						case ('2'):
 							if (tracksCount < 1) break;
 							resolve(1)
 							removeMsgAndCollector()
 							break
-						case (reactions[2]):
+
+						case ('3'):
 							if (tracksCount < 2) break;
 							resolve(2)
 							removeMsgAndCollector()
 							break;
-						case (reactions[3]):
+
+						case ('4'):
 							if (tracksCount < 3) break;
 							resolve(3)
 							removeMsgAndCollector()
 							break;
-						case (reactions[4]):
+						case ('5'):
 							if (tracksCount < 4) break;
 							resolve(4)
 							removeMsgAndCollector()
 							break;
-						case (cancel):
+						case ('c' || 'cancel' || 'с'):
 							resolve(-1)
 							removeMsgAndCollector()
 							break
 					}
 				})
-				for (let i = 0; i < tracksCount; i++) if (!chooseMsg.deleted) await chooseMsg.react(reactions[i])
-				if (!chooseMsg.deleted) await chooseMsg.react(cancel);
-			
-				async function removeMsgAndCollector() {
-					try {
-						clearTimeout(timeoutId);
-						reactionCollector.stop();
-						if (!chooseMsg.deleted) {
-							await chooseMsg.delete();
-						}
-					} catch (e) {
-						this.log('Remove choose msg error')
-					}
-				}
+				
 			} catch (e) {
 				this.log('Choose track error:', e)
 			}
 		})
 }
 	async sendControllMessage(channel: TextChannel, guildId: string) {
+		const msg = this.playerControllMessages.get(guildId);
+		if (msg) return;
 		if ((await channel.messages.fetch({}, {
 			cache: true
 		})).size > 0) await channel.bulkDelete((await channel.messages.fetch({}, {
@@ -306,23 +351,33 @@ export class Player extends BaseModule {
 	}
 	async addToQueue(message: Message) {
 		try {
+			if (!message.member.voice.channel) {
+				const msg = await message.channel.send({
+					embeds: [{ "description": "❌ **Вы должны находиться в голосовом канале!**", "color": 8340425, }]
+				});
+				this.client.utils.deleteMessageTimeout(msg, 5000);
+				return;
+			}
 			const guildId = message.guildId
 			let search = await this.searchTrack(message.content)
 			if (!search || search.length == 0) {
+				//https://www.youtube.com/watch?v=fmI_Ndrxy14&list=PL3z7nJoxQbExopK76qn2032OBBJ0F_4F_&index=2
+				if (message.content.match(/.*(?:(?:youtu.be\/)|(?:v\/)|(?:\/u\/\w\/)|(?:embed\/)|(?:watch\?))\??v?=?([^#\&\?]*).*/) && ytdl.validate(message.content)) {
+					
+				} 
 				const msg = await message.channel.send('Трек не найден.')
 				this.client.utils.deleteMessageTimeout(msg, 5000);
-				this.client.utils.deleteMessageTimeout(message, 5000);
 				return false;
 			};
-			if (!message.deleted) {
-				await message.delete()
-			}
+			this.blockedUsers.set(message.member.id, true)
 			let selectedTrack = await this.sendChooseMessage(message.member, <TextChannel>message.channel, search)
 			if (selectedTrack == -1) return;
+			const duration = (await ytdl.video_basic_info(search[selectedTrack].link)).video_details.durationRaw
 			let song = {
 				'title': search[selectedTrack].title,
 				'link': search[selectedTrack].link,
-				'repeat': false
+				'repeat': false,
+				'duration': duration
 			}
 		
 			let queue: Queue = {}
@@ -348,8 +403,8 @@ export class Player extends BaseModule {
 	async updateControllMessage(guildId: string) {
 		const msg = this.playerControllMessages.get(guildId);
 		if (!msg) return;
-		let newEmbed = Object.assign({}, embedTemplate)
-		let queue = db.getObject<Queue>('/queue')[guildId];
+		let newEmbed = JSON.parse(JSON.stringify(embedTemplate));
+		let queue = await this.getQueue(guildId);
 		if (queue.length == 0) {
 			const newMsg = await msg.edit({
 				embeds: [newEmbed]
@@ -360,12 +415,12 @@ export class Player extends BaseModule {
 		for (let i = 0; i < queue.length; i++) {
 			if (i < 11) {
 				if (i === 0) {
-					if (queue[i].repeat) newEmbed.title = `🔁 ${queue[i].title}`;
-					else newEmbed.title = `${queue[i].title}`;
+					if (queue[i].repeat) newEmbed.title = `🔁 [${queue[i].duration}] ${queue[i].title}`;
+					else newEmbed.title = `[${queue[0].duration}] ${queue[i].title}`;
 					newEmbed.url = queue[i].link;
 					continue;
 				}
-				newEmbed.description += `${i}. **[${queue[i].title}](${queue[i].link})** \n`;
+				newEmbed.description += `${i}. **[[${queue[i].duration}] ${queue[i].title}](${queue[i].link})** \n`;
 			} else {
 				newEmbed.fields[0] = {
 					"name": "\u200B",
@@ -375,7 +430,7 @@ export class Player extends BaseModule {
 				newEmbed.fields[0].value = ` ...Еще ${(queue.length - (i))}`
 				break
 			}
-
+		
 		}
 		const newMsg = await msg.edit({
 			embeds: [newEmbed]
@@ -409,49 +464,75 @@ export class Player extends BaseModule {
 		}
 	}
 	async searchAndPlayOrAddToQueue(message: Message) {
-		if (message.channelId !== this.channels.get(message.guildId)) return;
-		const search = await this.addToQueue(message);
-		if (!search) return;
-		if (this.player.has(message.guildId)) {
-			const player = this.player.get(message.guildId)
-			if (player.state.status == AudioPlayerStatus.Playing || player.state.status == AudioPlayerStatus.Paused) return;
+		try {
+			if (message.channelId !== this.channels.get(message.guildId)) return;
+			if((await this.client.checkBlackListUser(message.author.id)) !== null) {
+				const msg = await message.channel.send(`<@${message.author.id}> Вы внесены в черный список и не можете использовать команды этого бота. \n Причина: ${await this.client.checkBlackListUser(message.author.id)}`);
+				this.client.utils.deleteMessageTimeout(msg, 10000)
+				message.delete()
+				return
 		}
+			if (this.blockedUsers.get(message.member.id) == true) return;
+			this.client.utils.deleteMessageTimeout(message, 1000)
+			const search = await this.addToQueue(message);
+			if (!search) return;
+			if (this.player.has(message.guildId)) {
+				const player = this.player.get(message.guildId)
+				if (player.state.status == AudioPlayerStatus.Playing || player.state.status == AudioPlayerStatus.Paused) return;
+			}
 			
-		let connection = getVoiceConnection(message.guild.id)
-		if (!connection) {
-			 connection = joinVoiceChannel({
-				channelId: message.member.voice.channel.id,
-				guildId: message.guild.id,
-				adapterCreator: message.guild.voiceAdapterCreator
-			})
-		}
-		await this.play(message.guildId, connection)
+			let connection = getVoiceConnection(message.guild.id)
+			if (!connection) {
+				if (!message.member.voice.channel) {
+					const msg = await message.channel.send({
+						embeds: [{ "description": "❌ **Вы должны находиться в голосовом канале!**", "color": 8340425, }]
+					});
+					this.client.utils.deleteMessageTimeout(msg, 5000);
+					return;
+				}
+				connection = joinVoiceChannel({
+					channelId: message.member.voice.channel.id,
+					guildId: message.guild.id,
+					adapterCreator: message.guild.voiceAdapterCreator
+				})
+			}
+			await this.play(message.guildId, connection)
 		
-
+		} catch (e) {
+			this.log('Search of add to queue and play error:', e)
+		}
 	};
 	async play(guildId: string, connection?: VoiceConnection) {
 		if (!connection) connection = getVoiceConnection(guildId)
 			
 		if (!connection) return;
 		 
-			
+		
 		const player = await this.createPlayer(guildId)
+		
 		if (player.state.status == AudioPlayerStatus.Paused) {
 			connection.subscribe(player)
 			player.unpause()
 			return;
 		}
 		
-		if (!db.exists('/queue')) return;
-		let queue = db.getObject<Queue>('/queue')[guildId]
-		if (queue.length == 0) return;
-		const stream = await ytdl.stream(queue[0].link, process.env.YOUTUBECOOKIES);
-		const resource = createAudioResource(stream, { inputType: StreamType.Arbitrary });
 		
+		let queue = await this.getQueue(guildId)
+	
+		if (queue.length == 0) return;
+		// ytdl.setToken({
+		// 	youtube: {
+		// 		cookie: ""
+		// 	}
+		// })
+		const stream = await ytdl.stream(queue[0].link);
+	
+		const resource = createAudioResource(stream.stream, { inputType: StreamType.Arbitrary });
+
 		player.play(resource);
 		connection.subscribe(this.player.get(guildId))
 		player.unpause()
-		
+
 		
 	}
 	async createPlayer(guildId: string) {
@@ -463,17 +544,22 @@ export class Player extends BaseModule {
 			})
 			
 			this.player.set(guildId, player);
+			
+		
 			this.player.get(guildId).on(AudioPlayerStatus.Idle, async () => {
 				let connection = getVoiceConnection(guildId)
 				if (!connection) {
 					this.player.get(guildId).pause(true);
+					
 				}
-				let queue = db.getObject<Queue>('/queue')[guildId]
+				let queue = await this.getQueue(guildId)
+			
 				if (queue.length == 0) {
+					
 					connection.destroy();
 					return;
 				};
-				
+			
 				await this.nextSong(guildId)
